[Java.Interop.Dynamic-Tests] Fix WTF comment.
The cause of the "WTF" comment was that I was accessing an `int`
field, yet we were getting "java/lang/Integer" instead of "I" as the
JNI type reference.

The cause for this is that the wrong input was used;
JavaVM.GetJniTypeInfoForType(fieldType) should have been used, not
JavaVM.GetJniTypeInfoForJniTypeReference(JniClassName).

(The perils of not enough sleep!)