[Java.Interop.Dynamic] Don't call GetStaticFieldValue() twice.
An oddity appears when reading all the debug spew in
Java.Interop.Dynamic and the unit tests:
DynamicJavaClass.GetStaticFieldValue() was being invoked *twice* for
each single access.

Reading a static field:

	int max = Integer.MAX_VALUE

would result in:

	# MyMetaObject..ctor: paramter=$arg0 System.Linq.Expressions.ParameterExpression System.Object
	# MyMetaObject..ctor: value=Java.Interop.Dynamic.DynamicJavaClass Java.Interop.Dynamic.DynamicJavaClass
	GetMember: Expression=$arg0 [System.Object]; property=MAX_VALUE
	# FieldAccessInfo.GetMetaObject: parameter=$arg0
	MyHelperObject: e=$arg0
	MetaStaticFieldAccessObject.Convert: Expression=$arg0 [System.Object]
	# DynamicJavaClass(java/lang/Integer).field(MAX_VALUE) as System.Int32
	# DynamicJavaClass(java/lang/Integer).field(MAX_VALUE) as System.Int32

Note that there are two "field(MAX_VALUE) as System.Int32" messages;
each of these comes from a separate
DynamicJavaClass.GetStaticFieldValue() method invocation.

That's clearly crazy, but why does this happen?

The cause (lack of knowledge about DLR use! ;-) was that
MetaStaticFieldAccessObject.BindConvert() generated an Expression to
invoke DynamicJavaClass.GetStaticFieldValue():

	var field       = Expression.Convert(Expression, typeof(StaticFieldAccess));
	var instance    = Expression.Property (field, "JavaClass");
	var fieldName   = Expression.Property (field, "FieldName");
	var gsfv        = typeof(DynamicJavaClass).GetMethod ("GetStaticFieldValue");
	var expr        = Expression.Convert (
			Expression.Call (instance, gsfv, fieldName, Expression.Constant (binder.Type)),
			binder.Type);

This is all fine and good; the problem was that `expr` was *reused*
in the BindingRestrictions invocation:

	BindingRestrictions.GetTypeRestriction (expr, binder.Type);

It seems that in order for BindingRestrictions to be evaluated at
runtime, the expression provided is evaluated. This is why
DynamicJavaClass.GetStaticFieldValue() is invoked twice: `expr` is
evaluated twice, once to compute the value, and once again as part of
the BindingRestrictions logic to see if the value can be cached.

Using BindingRestrictions.GetInstanceRestriction() instead allows
`expr` to be evaluated only once, which in turn means
DynamicJavaClass.GetStaticFieldValue() is only evaluated once.

(I have no idea how to actually unit test this short of reading the
current diagnostic messages, which will need to die at some point...)