[Android.Interop-Tests] Initial Performance Tests.
So now that we're able to deploy and run on an Android device, how bad
is Java.Interop, performance-wise?

To check this, create a java.lang.Object(), and call Object.toString()
on the instance for 1000 times.

Results?

	I/mono-stdout( 3889):    Java.Interop Object.toString() Timing: 00:00:02.8221896
	I/mono-stdout( 3889): Xamarin.Android Object.toString() Timing: 00:00:00.0770231

Xamarin.Android invokes the method ~36x faster than Java.Interop.

This is clearly unacceptable. This needs to be profiled to determine
why Java.Interop is so much slower, and what can be done to speed
things up. :-(