[Interop-Tests] Use a JavaVMFixture type to centralize JVM creation.
Commit 3096aceb had the right idea -- There Can Be Only One JVM™ --
but the wrong implementation.

The problem with using a JVM.Current instance which all other tests
used is that it requires that everything use JVM.Current to ensure
it's initialized, instead of the preferred "normal" idiom of using
JavaVM.Current directly. ("Normal" meaning "that's what I expect
actual users to do, and that's what Java.Interop.Export does.")

Another problem is it was very easy to overlook accessing the
instance, making tests prone to ordering issues, and also resulting in
a failure within Android-Tests [0].

Instead, introduce a new JavaVMFixture type, which is to be used as
the base class for all NUnit Test Fixtures that require access to
JavaVM.Current. The JavaVMFixture contains a static constructor which
is responsible for initializing the JVM (via a partial method, so that
it may be separately implemented for Desktop & Android profiles).

This forced inheritance will ensure that the JavaVMFixture static
constructor is executed as part of test fixture creation, which in
turn will ensure that JavaVM.Current is valid, and allow us to cleanup
many of the `#pragma warning disables`s running around.

[0]:
INSTRUMENTATION_RESULT: failure: Android.InteropTests.TestsSample.CompareJniInvocationTiming=System.InvalidOperationException : No JavaVM has been created. Please use Java.Interop.JreVMBuilder.CreateJreVM().
  at Java.Interop.JavaVM.get_Current () [0x00000] in <filename unknown>:0
  at Java.Interop.JniEnvironment.<RootEnvironments>m__0 () [0x00000] in <filename unknown>:0
  at System.Threading.ThreadLocal`1[Java.Interop.JniEnvironment].GetSlowPath () [0x00000] in <filename unknown>:0