[PerformanceTests] Split up Java.Interop-PerformanceTests
Java.Interop-PerformanceTests should join the fun, a'la
commit 577587c0, splitting out a PerformanceTests Shared Project from
the existing Java.Interop-PerformanceTests desktop project.

This is so that Android.Interop-Tests can easily reuse the
PerformanceTests shared project in a future commit.

(Android.Interop-Tests isn't using PerformanceTests in this commit
because it causes ART to abort with too many JNI Local References.)