[Java.Interop.Dynamic] Turn into a PCL library, use on Android.
Java.Interop.Dynamic was originally written as a Desktop profile
library for two reasons:

 1. It "needed" Mono.Linq.Expressions to assist debugging, and
    Mono.Linq.Expressions wasn't a PCL; and

 2. JniPeerMembers required that the Type passed follow certain
    requirements, specifically that the type have a [JniTypeInfo]
    custom attribute, which in turn requires a TypeBuilder, which
    isn't provided in any PCL profile.

I *really* want Java.Inteorp.Dynamic as a PCL and on Android, so it's
time to address these.

To fix (1), just remove use of Mono.Linq.Expressions; it's only
needed for debugging. (I should also look into turning
Mono.Linq.Expressions into a PCL as well...)

To fix (2)...punt: add a new private
JniPeerMembers.CreatePeerMembers() method which skips the
Type/[JniTypeInfo] check, use Reflection to invoke
CreatePeerMembers(), and "fix" JniPeerInstanceMethods so that when
DeclaringType is null -- which it will be if DynamicJavaClass created
the instance -- we treat it the same as if self.GetType() matched
DeclaringType:

	// old and busted
	if (self.GetType () == DeclaringType)
		...

	// New hotness
	if (self.GetType () == DeclaringType || DeclaringType == null)
		...

(The JniPeerInstanceMethods change is required for Performance-Tests
to pass, including Android.Interop-Tests.)

Finally, turn the Java.Interop.Dynamic unit tests into a Shared
Project so that they may be easily used between the (new)
Java.Interop.Dynamic-Tests NUnit project and Android.Interop-Tests.