[Interop-Tests] Fix a TearDown failure.
I just recently noticed this, and I'm not sure when it started, but
commit 959e14cf was causing a TearDown failure...even though all tests
would pass [0]!

	Errors and Failures:
	1) TearDown Failure : Java.InteropTests.JavaObjectArray_object_ContractTest
	     JNI global references: grefStartCount=12; gref=13
	  Expected: True
	  But was:  False

I'm not sure why a TearDown failure doesn't count as a "failure" --
note that the message states "SUCCESS" -- but there is an error here,
and it shouldn't exist.

Oddly, this error looks like it was present when
JavaObjectArray_object_ContractTest.EndCheckGlobalRefCount() was first
added!

Regardless, the cause of the error is because creating the JavaObject
instance causes a GREF to be allocated for
JavaObject._members.JniPeerType (JniPeerMembers.JniPeerType), and
since JavaObject._members is `static` we can't expect this GREF to be
freed until the process exits.

Consequently, this isn't a "leak" -- at least, it's not a leak that
the test is supposed to be checking for.

Remove the JavaObject instance allocation, which will free up the GREF
allocation and remove the TearDown failure.

[0]:
> ############################################################################
> ##############                 S U C C E S S               #################
> ############################################################################
> Executed tests       : 319
> Ignored tests        : 0
> Failed tests         : 0
> Unhandled exceptions : 0
> Total time           : 6.6495005 seconds
> ############################################################################
>
> Tests run: 319, Errors: 0, Failures: 0, Inconclusive: 0, Time: 6.6495005 seconds
>   Not run: 0, Invalid: 0, Ignored: 0, Skipped: 0
>
> Errors and Failures:
> 1) TearDown Failure : Java.InteropTests.JavaObjectArray_object_ContractTest
>      JNI global references: grefStartCount=12; gref=13
>   Expected: True
>   But was:  False
>
> at Java.InteropTests.JavaObjectArray_object_ContractTest.EndCheckGlobalRefCount () [0x00056] in /Users/Shared/Dropbox/Developer/Java.Interop/src/Java.Interop/Tests/Java.Interop/JavaObjectArrayTest.cs:148