[Java.Interop] Rename IJavaObject to IJavaPeerable
Fixes: https://github.com/xamarin/Java.Interop/issues/21

Xamarin.Android has an Android.Runtime.IJavaObject interface type.
While Java.Interop.IJavaObject is in a different namespace, having
identical type names still means you'd need to fully qualify or use a
`using-alias`, which is annoying.

Rename IJavaObject to IJavaPeerable, which is also somewhat in keeping
with commit 25de1f38 which renamed IJavaObject.SafeHandle to
IJavaObject.PeerReference. Now, the property is
IJavaPeerable.PeerReference, further emphasizing the nature of what
the interface represents: a wrapper of a Java-side "peer" instance.

Related (for terminology):

	http://developer.xamarin.com/guides/android/advanced_topics/garbage_collection/

Modified to be consistent with this commit/name change:

> **Peer objects**: types which implement IJavaPeerable, e.g. all
> JavaObject and JavaException subclasses. Instances of these types
> have two "halfs" a managed peer and a native peer. The managed peer
> is an instance of the C# class. The native peer is an instance of a
> Java class within JVM, and the C# IJavaPeerable.PeerReference
> property contains a JNI object reference to the native peer.