[jnienv-gen, Java.Interop] Split out object reference #defines.
Previously, we used "one" #define to control both JniObjectReference
and JniEnvironment behavior: FEATURE_HANDLES_ARE_SAFE_HANDLES for
SafeHandle use, and FEATURE_HANDLES_ARE_INTPTRS for IntPtr use.

The problem is that  with commit 926e4bc4, there is no longer a 1:1
mapping between JniEnvironment and JniObjectReference behavior:
JniObjectReference + SafeHandles is the same, but JniObjectReference
and IntPtrs is supposed by both the JIIntPtrs and JIPinvokes
JniEnvironment implementations.

To better support this, separate out JniEnvironment and
JniObjectReference behavior.

JniEnvironment behavior is controlled by one of:

  * FEATURE_JNIENVIRONMENT_JI_INTPTRS
  * FEATURE_JNIENVIRONMENT_JI_PINVOKES
  * FEATURE_JNIENVIRONMENT_SAFEHANDLES
  * FEATURE_JNIENVIRONMENT_XA_INTPTRS

JniObjectReference is controlled by one of:

  * FEATURE_JNIOBJECTREFERENCE_INTPTRS
  * FEATURE_JNIOBJECTREFERENCE_SAFEHANDLES