[Java.Interop] Remove JavaObject/etc. from XAIntegrationDebug.
As per commit 25de1f38, we want make the XAIntegrationDebug
configuration (commit 54a2aff0) a *subset* of the full
Java.Interop.dll so that we can expedite inclusion of Java.Interop
within Xamarin.Android.

Commit 54a2aff0 created the configuration. THIS commit moves various
files into an <ItemGroup/> which excludes the XAIntegration
configuration, and adds #if-voodoo to other files to allow a subset of
Java.Interop to actually compile.

Meaning `make xa-all` actually builds a smaller assembly now:

	$ make
	$ make xa-all
	$ ls -l bin/*/Java.Interop.dll
	-rwxr-xr-x@ 1 jon  staff  535552 Oct 26 15:40 bin/Debug/Java.Interop.dll
	-rwxr-xr-x@ 1 jon  staff  142336 Oct 26 14:50 bin/XAIntegrationDebug/Java.Interop.dll

The resulting subset, at ~142K, is ~1/4 the size of the "full" lib.

NOTE: Commit 25de1f38 is BROKEN. It *builds* (on mono 4.0.4), but is
BROKEN, as shown in [0].

The problem is that mcs had a compiler bug wherein it would allow
invalid code to successfully compile, based on the ordering of files
on the build command line.

Which brings us to 25de1f38: that code shouldn't compile, but it did,
because of the mcs bug.

It's because of this that Java.Interop.csproj has *three* <ItemGroup/>
sections:

	<ItemGroup EVERYTHING />
	<ItemGroup Not-Xamarin.Android-Inetegration />
	<ItemGroup EVERYTHING />

There should be *two*, as in 54a2aff0; the final group is needed to
ensure that JniObjectReference.cs is *last*, allowing it to compile.

But don't forget that 25de1f38 and on -- including this commit! --
never should have built in the first place!

Fixing this will be done in a later commit.

[0]: https://bugzilla.xamarin.com/show_bug.cgi?id=35269