[Java.Interop] Rename JniTypeInfo to JniTypeSignature; make immutable.
I didn't like the name `JniTypeInfo`, because...reaons? Anyway, didn't
like it. Trying JniTypeSignature as a replacement.

Additionally, JniTypeInfo was a mutable struct, and mutable structs
are EVIL INCARNATE, so make JniTypeSignature *immutable*.

Add JniTypeSignature.GetPrimitiveWrapper(), so we can easily get the
JniTypeSignature for a primitive wrapper, e.g. the primitive wrapper
for `I` is `java/lang/Integer`.

Stop using JniTypeSignature.ToString() as *the* way to get a JNI Type
Reference for use with JniEnvironment.Types.FindClass()/etc., and
instead add a new JniTypeSignature.Name property that serves the same
purpose. This should be clearer to the reader.

JniTypeSignature.ToString(), meanwhile, is changed to be *never*
useful with JNI signatures, to help catch bugs with the introduction
and use of the Name property.

Finally, make JniTypeSignature implement IEquatable<JniTypeSignature>.
It's not currently needed, but since it's an immutable type...