[Java.Interop] Remove IJniObjectReferenceManager, use class.
I'm not entirely confident that the IJniObjectReferenceManager
abstraction was correct/complete, so to simplify possible future
versioning, turn this into a class with virtual members.

This allows LoggingJniObjectReferenceManagerDecorator to override and
delegate everything -- as it did before -- while allowing us to add
new virtual methods in the future without breaking everything.

	https://msdn.microsoft.com/en-us/library/vstudio/ms229013(v=vs.100).aspx
> **Do favor defining classes over interfaces.**
>
> In later versions of your library, you can safely add new members to
> classes; you cannot add members to interfaces without breaking
> existing code.