[Java.Interop] Make JniEnvironment `static`; add JniTransition.
Recall from commit 25de1f38:

>  1. @jassmith has been profiling Xamarin.Android, and GC allocations
>     would be "a problem", as increased allocations will increase
>     GC-related overheads, slowing down the app.
>
>     We want a "GC steady state" wherein once things are running,
>     and things are cached, future JNI invocations such as
>     JNIEnv::CallObjectMethod() won't allocate additional garbage
>     (unless *Java* is returning new "garbage" instances...)

We want a GC "steady state", meaning NO NEW OPTIONAL ALLOCATIONS.
("Required" allocations would include new wrappers for Java instances
that entered managed code for the first time.)

Which brings us to JniEnvironment, which has always been a
non-`static` reference type and in commit 2f9fece8 required an
allocation on every Java-to-managed transition:

	TReturnType (*)(IntPtr jnienv, IntPtr context /* ... args ....*/ )
	{
		var __envp = new JniEnvironment (jnienv);
		try {
			...
		}
		catch (Exception e) {
			__envp.SetPendingException (e);
		}
		finally {
			__envp.Dispose ();
		}
	}

That `new JniEnvironment` is required so that we have a "scope" around
which to dispose of allocated JniLocalReference instances, alongside
the previous requirements of checking the JNIEnv* pointer for
invalidation/etc (which JniEnvironment.CheckCurrent() did).

We *want* a data structure for Java-to-managed invocations.

We *don't* want a GC allocation.

Enter the JniTransition struct, which is a subset of JniEnvironment,
but a struct, so it implies no GC allocations:

	public struct JniTransition : IDisposable {

		public JniTransition (IntPtr environmentPointer);

		public void SetPendingException (Exception exception);
		public void Dispose ();
	}

Use is identical to the previous JniEnvironment:

	TReturnType (*)(IntPtr jnienv, IntPtr context /* ... args ....*/ )
	{
		var __envp = new JniTransition(jnienv);
		try {
			...
		}
		catch (Exception e) {
			__envp.SetPendingException (e);
		}
		finally {
			__envp.Dispose ();
		}
	}

The one conceptual problem is that JniTransition IS a mutable struct,
and mutable structs are EVIL (e6e13654), so why is this acceptable?
It's acceptable because NOBODY SHOULD EVER USE IT. It would be used by
a binding generator and the runtime marshal method generators; it
shouldn't need to be used by developers, much less *copied*, which is
where the real issues with mutable structs come in.

With JniTransition in place, turn JniEnvironment into a `static`
class. This has the unfortunate effect of breaking *everything*, but
it has the benefit that we're no longer creating JniEnvironmentInvoker
multiple times on a single thread (and probably re-caching the
delegates for the same method OVER AND OVER AND...).

As such, JniEnvironment.Current.JavaVM becomes JniEnvironment.Runtime
(I'm finding `JavaVM` more annoying the more I look at it), and the
other useful instance members become static members.

We also introduce JniEnvironmentInfo, which contains the per-thread
JniEnvironment information. Ideally this would be internal, but it
needs to be public so that JniObjectReferenceManager can use it for
JNI Local Reference accounting; *something* per-thread needs to be
updated for JNI local references...