[Java.Interop+XAIntegration] Delay-sign Java.Interop.dll.
Mono.Android.dll is signed and strong-named. Consequently, all of its
dependencies must ALSO be signed and strong-named, which would include
Java.Interop.dll.

Update Java.Interop.csproj so that the XAIntegrationDebug
configuration will delay-sign the assembly.

Update Makefile so that the `make xa-all` target now accepts an SNK
make variable, which is the location of the keyfile to use for
signing:

	make xa-all SNK=path/to/signing_key.snk