[Java.Interop] Move JniEnvironment.References.Dispose to JniObjectReference
This is to aid in IDE discoverability: JniEnvironment.References might
as well not exist; who's going to know to look there?

The user has a JniObjectReference value, and you can create *new*
values from the value, e.g. JniObjectReference.NewLocalRef(), so how
are you supposed to know to go to JniEnvironment.References to
*destroy* that value?

Answer: they won't.

Move JniEnvironment.References.Dispose() to
JniObjectReference.Dispose(), so that "normal" reference creation and
destruction are in the ~same place (except one is an instance method
while the other is a static method....).