[Java.Interop] Special-case return types.
Special-case the following methods:

  * JNIEnv::EnsureLocalCapacity()
  * JNIEnv::GetJavaVM()
  * JNIEnv::MonitorEnter()
  * JNIEnv::MonitorExit()
  * JNIEnv::PushLocalFrame()
  * JNIEnv::RegisterNatives()
  * JNIEnv::Throw()
  * JNIEnv::ThrowNew()
  * JNIEnv::UnregisterNatives()

JNIEnv::EnsureLocalCapacity() are special JNIEnv::PushLocalFrame() in
that they don't have a "THROWS" section in their [documentation][0],
but their their description states that an exception is thrown and/or
pending:

> **EnsureLocalCapacity**
> Returns 0 on success; otherwise returns a negative number and throws
> an OutOfMemoryError

Meaning I can't always rely on a THROWS documentation section to know
whether or not an exception may be pending. :-(

Meanwhile, JNIEnv::Throw() and JNIEnv::ThrowNew() document THROWS
sections -- they kinda HAVE TO ;-) --  but they only throw if the
method returns a non-zero value, meaning the public API *should*
have a `void` return type, with the exception handled internally.

JNIEnv::MonitorEnter() is "reasonable" enough: non-zero value means
"failure" -- but no defintion of what "failure" *means* (pending
exception?) -- so we'll map "failure" to an InvalidOperationException.
JNIEnv::MonitorExit(), meanwhile, isn't quite so reasonable, in that
the docs mention an IllegalMonitorStateException, but don't say that a
pending exception is the *only* reason MonitorExit() may fail.

Failures in JNIEnv::GetJavaVM(), JNIEnv::RegisterNatives(), and
JNIEnv::UnregisterNatives() are mapped to InvalidOperationException.

[0]: http://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/functions.html