[Java.Interop] API Review: Add JniReleaseArrayElementsMode
The [JNIEnv `Release<PrimitiveType>ArrayElements` Routines][0] methods
have a `jint mode` parameter which can have one of only three values:
0, JNI_COMMIT, and JNI_ABORT.

The binding for these methods was using just a plain `int`, leaving it
up to the caller to "know" the value to pass for this parameter:

	JniEnvironment.Arrays.ReleaseBooleanArrayElements (
			array,
			elements,
			??? uh, what value do we use here? Hit the docs!
	);

Improve matters by binding the `mode` parameter as the new
JniReleaseArrayElementsMode enumeration type:

	enum JniReleaseArrayElementsMode {
		Default,
		Commit,
		Abort,
	}

I'm not entirely fond of these names; for example, the "previous"
(removed) JniArrayElementsReleaseMode enum tried to use more
descriptive names:

	enum JniArrayElementsReleaseMode {
		CopyBack      = 0,
		DoNotCopyBack = 2,
	}

Notably absent was JNI_COMMIT, and I couldn't think of a good name for
it (CopyBackWithoutFree?).

([Naming is hard. Let's go shopping][1].)

Furthermore, in some scenarios the names are actually misleading; if
you call e.g. JNIEnv::GetBooleanArrayElements() and the `isCopy`
output value is `false` -- i.e. you have direct access to the
underlying array! -- then the `mode` value is *meaningless*: you've
been directly modifying the underlying array. There *is* no copy to be
done. Additionally, there's also no way to "abort".

Which really means this API is fucked. (Ugh.)

So the consensus was to KISS with the names: there's likely no way to
have a fully unambiguous name that describes all semantics, so a
"vague" name that is suggestive of the documentation might be the best
that we can do, hence the Default, Commit, and Abort values.

[0]: http://docs.oracle.com/javase/8/docs/technotes/guides/jni/spec/functions.html#Release_PrimitiveType_ArrayElements_routines
[1]: https://xamarinhq.slack.com/archives/android/p1448036303001867