[Java.Interop] Fix JniObjectReference + SafeHandle implementation
Commit e8cead2e noted a [compiler bug][0] that permitted invalid code
to compile, and that it should be fixed:

> The problem is that mcs had a compiler bug wherein it would allow
> invalid code to successfully compile, based on the ordering of files
> on the build command line.
> ...
> Fixing this will be done in a later commit.

It's now time to fix this. :-)

The problem is that JniObjectReference contains a reference type --
JniReferenceSafeHandle -- when FEATURE_JNIOBJECTREFERENCE_SAFEHANDLES
is #defined. This prevents using `JniObjectReference*`, as unsafe
pointers may not be used with structs that contain reference types.

The fix is to remove the JniReferenceSafeHandle member, which is where
System.Runtime.InteropServices.GCHandle comes in: instead of referring
directly to the JniReferenceSafeHandle instance, JniObjectReference
now contains a GCHandle to the JniReferenceSafeHandle.

Otherwise, there's no major changes here, but fixing the underlying
source of the compiler error allows us to fix Java.Interop.csproj to
not need the multiple <ItemGroup/> sections, and to declare "all" of
the JniObjectReferenceOptions members in the same source file, using
partial classes, instead of spreading them out between multiple files.
(Trying to use partial classes also raised the CS0208 error that is
fixed here, so using partial classes couldn't be done earlier.)

[0]: https://bugzilla.xamarin.com/show_bug.cgi?id=35269